requirements.txt
pandas==2.2.2
numpy==1.26.4
xgboost==2.0.3
scikit-learn==1.4.2
boto3==1.34.34
pyyaml==6.0.1
openpyxl==3.1.2
